backend/requirements.txt
Django>=3.0,<4.0
# psycopg2
# psycopg2-binary>=2.8
django-versatileimagefield
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
requests



djoser
environs
gunicorn
# psycopg2-binary
# psycopg2

